Injector/requirements.txt
requests~=2.28.2
kafka-python~=2.0.2